Raise an error on unauthorized client delete attempt
With central clients, users may attempt to issue delete request towards
them.

While deletion will not happen, as user_id/project_id match will not
happen,
API will reply with success code, which is confusing.

Let's properly raise an unauthorized in such case.

+++ freezer_api/db/sqlalchemy/api.py
@@ -506,7 +506,14 @@ def delete_client(user_id, client_id, project_id=None):
                           client_id=client_id)
 
     if existing:
-        clientt = existing[0].get('client')
+        client_info = existing[0]
+        if client_info.get('user_id') != user_id or \
+                client_info.get('project_id') != project_id:
+            raise freezer_api_exc.AccessForbidden(
+                "You are not permitted to delete this client"
+            )
+
+        clientt = client_info.get('client')
         clientid = clientt.get('uuid')
         delete_tuple(tablename=models.Client, user_id=user_id,
                      tuple_id=clientid, project_id=project_id)
